Split race discovery helpers
