+++ Shared.Tests/HtmlTextTests.cs
@@ -8,8 +8,8 @@ public class HtmlTextTests
     [InlineData(null, "")]
     [InlineData("", "")]
     [InlineData("  ", "")]
-    [InlineData("&lt;strong&gt;Ultra&lt;/strong&gt;", "<strong>Ultra</strong>")]
-    [InlineData("<p>Eco&nbsp;Trail</p><script>ignore()</script>", "Eco Trail ignore()")]
+    [InlineData("&lt;strong&gt;Ultra&lt;/strong&gt;", "Ultra")]
+    [InlineData("<p>Eco&nbsp;Trail</p><script>ignore()</script>", "Eco\u00a0Trail  ignore()")]
     public void DecodeAndStripTags_ReturnsPlainDecodedText(string? html, string expected)
     {
         Assert.Equal(expected, HtmlText.DecodeAndStripTags(html));
